+++ app/src/main/res/layout/activity_cat_list.xml
@@ -9,10 +9,6 @@
         android:name="com.nikhil.mybestfriend.feature.cat.listing.view.CatListFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginBottom="8dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
