+++ app/src/main/res/layout/activity_settings.xml
@@ -2,6 +2,7 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
+    android:background="@drawable/card_background"
     android:layout_height="match_parent">
 
     <androidx.fragment.app.FragmentContainerView
