+++ app/src/main/res/values/colors.xml
@@ -8,4 +8,5 @@
     <color name="defaultTitleColor">#4c4c4d</color>
     <color name="defaultTextColor">#4f4f50</color>
     <color name="white">#ffffff</color>
+    <color name="black">#000000</color>
 </resources>
