+++ app/src/main/res/values/styles.xml
@@ -25,8 +25,16 @@
         <item name="android:textStyle">bold</item>
     </style>
 
+    <style name="WhiteTitle" parent="Title">
+        <item name="android:textColor">@color/white</item>
+    </style>
+
     <style name="Text" parent="AppTheme">
         <item name="android:textSize">14sp</item>
+    </style>
+
+    <style name="WhiteText" parent="Text">
+        <item name="android:textColor">@color/white</item>
     </style>
 
     <style name="AppTheme.NoActionBar">
